+++ network/get_peers.js
@@ -19,6 +19,7 @@ const {returnResult} = require('asyncjs-util');
 const size = require('window-size');
 
 const {authenticatedLnd} = require('./../lnd');
+const {chartAliasForPeer} = require('./../display');
 const {formatFeeRate} = require('./../display');
 const {getPastForwards} = require('./../routing');
 const {sortBy} = require('./../arrays');
@@ -27,14 +28,12 @@ const defaultInvoicesLimit = 100;
 const defaultSort = 'first_connected';
 const fromNow = epoch => !epoch ? undefined : moment(epoch * 1e3).fromNow();
 const {isArray} = Array;
-const isEmoji = /(?:[\u261D\u26F9\u270A-\u270D]|\uD83C[\uDF85\uDFC2-\uDFC4\uDFC7\uDFCA-\uDFCC]|\uD83D[\uDC42\uDC43\uDC46-\uDC50\uDC66-\uDC69\uDC6E\uDC70-\uDC78\uDC7C\uDC81-\uDC83\uDC85-\uDC87\uDCAA\uDD74\uDD75\uDD7A\uDD90\uDD95\uDD96\uDE45-\uDE47\uDE4B-\uDE4F\uDEA3\uDEB4-\uDEB6\uDEC0\uDECC]|\uD83E[\uDD18-\uDD1C\uDD1E\uDD1F\uDD26\uDD30-\uDD39\uDD3D\uDD3E\uDDD1-\uDDDD])(?:\uD83C[\uDFFB-\uDFFF])?|(?:[\u231A\u231B\u23E9-\u23EC\u23F0\u23F3\u25FD\u25FE\u2614\u2615\u2648-\u2653\u267F\u2693\u26A1\u26AA\u26AB\u26BD\u26BE\u26C4\u26C5\u26CE\u26D4\u26EA\u26F2\u26F3\u26F5\u26FA\u26FD\u2705\u270A\u270B\u2728\u274C\u274E\u2753-\u2755\u2757\u2795-\u2797\u27B0\u27BF\u2B1B\u2B1C\u2B50\u2B55]|\uD83C[\uDC04\uDCCF\uDD8E\uDD91-\uDD9A\uDDE6-\uDDFF\uDE01\uDE1A\uDE2F\uDE32-\uDE36\uDE38-\uDE3A\uDE50\uDE51\uDF00-\uDF20\uDF2D-\uDF35\uDF37-\uDF7C\uDF7E-\uDF93\uDFA0-\uDFCA\uDFCF-\uDFD3\uDFE0-\uDFF0\uDFF4\uDFF8-\uDFFF]|\uD83D[\uDC00-\uDC3E\uDC40\uDC42-\uDCFC\uDCFF-\uDD3D\uDD4B-\uDD4E\uDD50-\uDD67\uDD7A\uDD95\uDD96\uDDA4\uDDFB-\uDE4F\uDE80-\uDEC5\uDECC\uDED0-\uDED2\uDEEB\uDEEC\uDEF4-\uDEF8]|\uD83E[\uDD10-\uDD3A\uDD3C-\uDD3E\uDD40-\uDD45\uDD47-\uDD4C\uDD50-\uDD6B\uDD80-\uDD97\uDDC0\uDDD0-\uDDE6])|(?:[#\*0-9\xA9\xAE\u203C\u2049\u2122\u2139\u2194-\u2199\u21A9\u21AA\u231A\u231B\u2328\u23CF\u23E9-\u23F3\u23F8-\u23FA\u24C2\u25AA\u25AB\u25B6\u25C0\u25FB-\u25FE\u2600-\u2604\u260E\u2611\u2614\u2615\u2618\u261D\u2620\u2622\u2623\u2626\u262A\u262E\u262F\u2638-\u263A\u2640\u2642\u2648-\u2653\u2660\u2663\u2665\u2666\u2668\u267B\u267F\u2692-\u2697\u2699\u269B\u269C\u26A0\u26A1\u26AA\u26AB\u26B0\u26B1\u26BD\u26BE\u26C4\u26C5\u26C8\u26CE\u26CF\u26D1\u26D3\u26D4\u26E9\u26EA\u26F0-\u26F5\u26F7-\u26FA\u26FD\u2702\u2705\u2708-\u270D\u270F\u2712\u2714\u2716\u271D\u2721\u2728\u2733\u2734\u2744\u2747\u274C\u274E\u2753-\u2755\u2757\u2763\u2764\u2795-\u2797\u27A1\u27B0\u27BF\u2934\u2935\u2B05-\u2B07\u2B1B\u2B1C\u2B50\u2B55\u3030\u303D\u3297\u3299]|\uD83C[\uDC04\uDCCF\uDD70\uDD71\uDD7E\uDD7F\uDD8E\uDD91-\uDD9A\uDDE6-\uDDFF\uDE01\uDE02\uDE1A\uDE2F\uDE32-\uDE3A\uDE50\uDE51\uDF00-\uDF21\uDF24-\uDF93\uDF96\uDF97\uDF99-\uDF9B\uDF9E-\uDFF0\uDFF3-\uDFF5\uDFF7-\uDFFF]|\uD83D[\uDC00-\uDCFD\uDCFF-\uDD3D\uDD49-\uDD4E\uDD50-\uDD67\uDD6F\uDD70\uDD73-\uDD7A\uDD87\uDD8A-\uDD8D\uDD90\uDD95\uDD96\uDDA4\uDDA5\uDDA8\uDDB1\uDDB2\uDDBC\uDDC2-\uDDC4\uDDD1-\uDDD3\uDDDC-\uDDDE\uDDE1\uDDE3\uDDE8\uDDEF\uDDF3\uDDFA-\uDE4F\uDE80-\uDEC5\uDECB-\uDED2\uDEE0-\uDEE5\uDEE9\uDEEB\uDEEC\uDEF0\uDEF3-\uDEF8]|\uD83E[\uDD10-\uDD3A\uDD3C-\uDD3E\uDD40-\uDD45\uDD47-\uDD4C\uDD50-\uDD6B\uDD80-\uDD97\uDDC0\uDDD0-\uDDE6])\uFE0F/g;
 const {max} = Math;
 const minutesPerBlock = network => network === 'ltcmainnet' ? 10 / 4 : 10;
 const notNull = array => array.filter(n => n !== null);
 const notFoundIndex = -1;
-const offlineEmoji = '💀';
-const shortKey = key => key.substring(0, 16);
-const sumOf = arr => arr.reduce((sum, n) => sum + n);
+const {round} = Math;
+const sumOf = arr => arr.reduce((sum, n) => sum + n, Number());
 const uniq = arr => Array.from(new Set(arr));
 const wideSizeCols = 150;
 
@@ -344,6 +343,25 @@ module.exports = (args, cbk) => {
             return channel.partner_public_key === publicKey;
           });
 
+          // A thaw channel cannot be cooperatively closed until a block height
+          const hasThawChannel = activeChannels
+            .map(n => n.cooperative_close_delay_height)
+            .filter(n => !!n)
+            .find(n => n > wallet.current_block_height);
+
+          const hasHtlcChannel = activeChannels
+            .map(n => n.pending_payments.length)
+            .filter(n => !!n)
+            .find(n => !!n.id);
+
+          const uptime = sumOf(activeChannels
+            .filter(n => !!n.time_online)
+            .map(n => n.time_online));
+
+          const downtime = sumOf(activeChannels
+            .filter(n => !!n.time_offline)
+            .map(n => n.time_offline));
+
           const pendingChannels = getPending.pending_channels.filter(chan => {
             return !!chan.is_opening && chan.partner_public_key === publicKey;
           });
@@ -381,11 +399,14 @@ module.exports = (args, cbk) => {
 
           return {
             alias: node.alias,
-            fee_earnings: feeEarnings.reduce((sum, {fee}) => sum + fee, 0),
+            downtime_percentage: round(100 * (downtime / (downtime + uptime))),
+            fee_earnings: sumOf(feeEarnings.map(n => n.fee)),
             first_connected: moment().subtract(blocks * mpb, 'minutes').unix(),
             inbound_fee_rate: feeRate,
             inbound_liquidity: sumOf(channels.map(n => n.remote_balance)),
+            is_forwarding: hasHtlcChannel || undefined,
             is_offline: !peer || undefined,
+            is_thawing: hasThawChannel || undefined,
             last_activity: args.idle_days !== undefined ? lastActivity : null,
             outbound_liquidity: sumOf(channels.map(n => n.local_balance)),
             public_key: publicKey,
@@ -425,12 +446,15 @@ module.exports = (args, cbk) => {
 
               return {
                 alias: peer.alias,
+                downtime_percentage: peer.downtime_percentage,
                 fee_earnings: peer.fee_earnings,
                 first_connected: fromNow(peer.first_connected),
                 last_activity: fromNow(peer.last_activity),
                 inbound_fee_rate: formatFeeRate({rate}).display,
                 inbound_liquidity: peer.inbound_liquidity,
+                is_forwarding: peer.is_forwarding,
                 is_offline: peer.is_offline,
+                is_thawing: peer.is_thawing,
                 outbound_liquidity: peer.outbound_liquidity,
                 public_key: peer.public_key,
               };
@@ -445,11 +469,14 @@ module.exports = (args, cbk) => {
             peers: peers.peers.map(n => ({
               alias: n.alias,
               fee_earnings: n.fee_earnings || undefined,
+              downtime_percentage: n.downtime_percentage || undefined,
               first_connected: n.first_connected || undefined,
               last_activity: n.last_activity || undefined,
               inbound_fee_rate: n.inbound_fee_rate || undefined,
               inbound_liquidity: n.inbound_liquidity || undefined,
+              is_forwarding: n.is_forwarding || undefined,
               is_offline: n.is_offline || undefined,
+              is_thawing: n.is_thawing || undefined,
               outbound_liquidity: n.outbound_liquidity || undefined,
               public_key: n.public_key || undefined,
             })),
@@ -462,7 +489,6 @@ module.exports = (args, cbk) => {
           peers: peers.peers,
           rows: []
             .concat([notNull([
-              '',
               'Alias',
               'Inbound',
               'In Fee',
@@ -486,9 +512,17 @@ module.exports = (args, cbk) => {
                 tokens: peer.outbound_liquidity,
               });
 
+              const alias = chartAliasForPeer({
+                alias: peer.alias,
+                downtime_percentage: peer.downtime_percentage,
+                is_forwarding: peer.is_forwarding,
+                is_inactive: peer.is_offline,
+                is_thawing: peer.is_thawing,
+                public_key: peer.public_key,
+              });
+
               return notNull([
-                peer.is_offline ? offlineEmoji : ' ',
-                peer.alias.replace(isEmoji, '') || shortKey(peer.public_key),
+                alias.display,
                 inbound.display || ' ',
                 peer.inbound_fee_rate || ' ',
                 outbound.display || ' ',