+++ test/network/test_reconnect.js
@@ -72,6 +72,7 @@ const tests = [
             },
             "remote_pubkey": "010000000000000000000000000000000000000000000000000000000000000000",
             "static_remote_key": true,
+            "thaw_height": 0,
             "total_satoshis_received": "1",
             "total_satoshis_sent": "1",
             "unsettled_balance": "1",